CLI and visualization
